# optimize index_select performance on CPU with TensorIterator