--- a/aten/src/ATen/Declarations.cwrap
+++ b/aten/src/ATen/Declarations.cwrap
@@ -150,11 +150,12 @@
 ]]
 [[
   name: _th_index_select
-  cpu_bool: True
   cuda_bool: True
   cname: indexSelect
   variants:
     - function
+  backends:
+    - CUDA
   return: argument 0
   scalar_check: self.dim() == 0
   arguments: