@mingfeima 2019-12-02
The current implementation of `index_select` from TH module has two performance drawbacks:
1. when input is contiguous, it does parallelization; but with non-contiguous input it runs in sequential.
2. non-contiguous input relies on `copy_` which requires creating sub tensors for each row to be copied. This is too heavy for `Embedding` since `embedding_dim` is usually not a very big number.

With non-contiguous input, this PR brings **799x** speedup on single socket and **31x** speedup on single core on Xeon Skylake 6148.

Below shows the performance compare between **original** and **this pr**, the benchmark will compare from:
1. a range of input size
2. contiguous tensor and non-contiguous tensor
3. single socket and single core

benchmark code available at this [op_bench-py](https://github.com/mingfeima/op_bench-py), to reproduce, `./run.sh index_select.py`. Test machine is Xeon Skylake 6148 with 2*20 cores @2.40GHz

1. input size is [num_embeddings, embedding_dim]
2. Unit is **index/second** a.k.a the throughput, the higher the better

#### single socket run
input size | index size | contiguous | original | this pr | speedup
-- | -- | -- | -- | -- | --
[10000, 128] | 10000 | TRUE | 406.872 | 506.43 | 1.24
[100000, 128] | 10000 | TRUE | 376.868 | 469.22 | 1.25
[1000000, 128] | 100000 | TRUE | 29.124 | 33.91 | 1.16
[10000000, 128] | 1000000 | TRUE | 34.433 | 37.90 | 1.10
[10000, 128] | 10000 | FALSE | 0.606 | 484.25 | 799.09
[100000, 128] | 10000 | FALSE | 0.609 | 455.83 | 748.49
[1000000, 128] | 100000 | FALSE | 0.543 | 35.02 | 64.49
[10000000, 128] | 1000000 | FALSE | 0.543 | 38.35 | 70.63

#### single core run
input size | index size | contiguous | original | this pr | speedup
-- | -- | -- | -- | -- | --
[10000, 128] | 10000 | TRUE | 21.362 | 21.71 | 1.02
[100000, 128] | 10000 | TRUE | 18.575 | 18.71 | 1.01
[1000000, 128] | 100000 | TRUE | 4.658 | 4.62 | 0.99
[10000000, 128] | 1000000 | TRUE | 4.625 | 4.75 | 1.03
[10000, 128] | 10000 | FALSE | 0.655 | 20.30 | 30.99
[100000, 128] | 10000 | FALSE | 0.66 | 18.68 | 28.31
[1000000, 128] | 100000 | FALSE | 0.585 | 4.69 | 8.01
[10000000, 128] | 1000000 | FALSE | 0.586 | 4.69 | 8.00

attach the original log:
**original**
```bash
### using OMP_NUM_THREADS=20
input size: [10000, 128]; index size: [10000]; contiguous; thoughput: indexs/us = 406.872
input size: [100000, 128]; index size: [10000]; contiguous; thoughput: indexs/us = 376.868
input size: [1000000, 128]; index size: [100000]; contiguous; thoughput: indexs/us = 29.124
input size: [10000000, 128]; index size: [1000000]; contiguous; thoughput: indexs/us = 34.433
input size: [10000, 128]; index size: [10000]; non-contiguous; thoughput: indexs/us = 0.606
input size: [100000, 128]; index size: [10000]; non-contiguous; thoughput: indexs/us = 0.609
input size: [1000000, 128]; index size: [100000]; non-contiguous; thoughput: indexs/us = 0.543
input size: [10000000, 128]; index size: [1000000]; non-contiguous; thoughput: indexs/us = 0.543

### using OMP_NUM_THREADS=1
input size: [10000, 128]; index size: [10000]; contiguous; thoughput: indexs/us = 21.362
input size: [100000, 128]; index size: [10000]; contiguous; thoughput: indexs/us = 18.575
input size: [1000000, 128]; index size: [100000]; contiguous; thoughput: indexs/us = 4.658
input size: [10000000, 128]; index size: [1000000]; contiguous; thoughput: indexs/us = 4.625
input size: [10000, 128]; index size: [10000]; non-contiguous; thoughput: indexs/us = 0.655
input size: [100000, 128]; index size: [10000]; non-contiguous; thoughput: indexs/us = 0.660
input size: [1000000, 128]; index size: [100000]; non-contiguous; thoughput: indexs/us = 0.585
input size: [10000000, 128]; index size: [1000000]; non-contiguous; thoughput: indexs/us = 0.586
```

**this pr**
```bash
### using OMP_NUM_THREADS=20
input size: [10000, 128]; index size: [10000]; contiguous; thoughput: indexs/us = 506.430
input size: [100000, 128]; index size: [10000]; contiguous; thoughput: indexs/us = 469.219
input size: [1000000, 128]; index size: [100000]; contiguous; thoughput: indexs/us = 33.907
input size: [10000000, 128]; index size: [1000000]; contiguous; thoughput: indexs/us = 37.898
input size: [10000, 128]; index size: [10000]; non-contiguous; thoughput: indexs/us = 484.249
input size: [100000, 128]; index size: [10000]; non-contiguous; thoughput: indexs/us = 455.830
input size: [1000000, 128]; index size: [100000]; non-contiguous; thoughput: indexs/us = 35.017
input size: [10000000, 128]; index size: [1000000]; non-contiguous; thoughput: indexs/us = 38.354

### using OMP_NUM_THREADS=1
input size: [10000, 128]; index size: [10000]; contiguous; thoughput: indexs/us = 21.707
input size: [100000, 128]; index size: [10000]; contiguous; thoughput: indexs/us = 18.714
input size: [1000000, 128]; index size: [100000]; contiguous; thoughput: indexs/us = 4.616
input size: [10000000, 128]; index size: [1000000]; contiguous; thoughput: indexs/us = 4.750
input size: [10000, 128]; index size: [10000]; non-contiguous; thoughput: indexs/us = 20.300
input size: [100000, 128]; index size: [10000]; non-contiguous; thoughput: indexs/us = 18.684
input size: [1000000, 128]; index size: [100000]; non-contiguous; thoughput: indexs/us = 4.687
input size: [10000000, 128]; index size: [1000000]; non-contiguous; thoughput: indexs/us = 4.686
```