@Jianhui-Li 2019-12-03
@gottbrath @VitalyFedyunin  FYI index_select moved from TH to ATEN