@VitalyFedyunin 2019-12-05
Hi! I'm starting review now, meanwhile please remove old TH code.