@mingfeima 2019-12-06
old TH code has been removed.