@mingfeima 2019-12-20
@VitalyFedyunin please review the code!